Refactor integration
