+++ src/golang/lib/models/shared/integration_config.go
@@ -1,10 +1,9 @@
 package shared
 
 import (
-	"database/sql/driver"
 	"fmt"
 
-	"github.com/aqueducthq/aqueduct/lib/models/utils"
+	"github.com/aqueducthq/aqueduct/lib/collections/shared"
 	"github.com/dropbox/godropbox/errors"
 )
 
@@ -21,7 +20,7 @@ type S3IntegrationConfig struct {
 	Type              S3ConfigType `json:"type"`
 	Bucket            string       `json:"bucket"`
 	Region            string       `json:"region"`
-	AccessKeyID       string       `json:"access_key_id"`
+	AccessKeyId       string       `json:"access_key_id"`
 	SecretAccessKey   string       `json:"secret_access_key"`
 	ConfigFilePath    string       `json:"config_file_path"`
 	ConfigFileContent string       `json:"config_file_content"`
@@ -34,7 +33,7 @@ type AirflowIntegrationConfig struct{}
 
 // GCSIntegrationConfig contains the fields for connecting a Google Cloud Storage integration.
 type GCSIntegrationConfig struct {
-	GCSConfig
+	shared.GCSConfig
 	UseAsStorage ConfigBool `json:"use_as_storage"`
 }
 
@@ -42,6 +41,23 @@ type K8sIntegrationConfig struct {
 	KubeconfigPath string     `json:"kubeconfig_path" yaml:"kubeconfigPath"`
 	ClusterName    string     `json:"cluster_name"  yaml:"clusterName"`
 	UseSameCluster ConfigBool `json:"use_same_cluster"  yaml:"useSameCluster"`
+}
+
+type LambdaIntegrationConfig struct {
+	RoleArn string `json:"role_arn" yaml:"roleArn"`
+}
+
+type DatabricksIntegrationConfig struct {
+	// WorkspaceURL is the full url for the Databricks workspace that
+	// Aqueduct operators will run on.
+	WorkspaceURL string `json:"workspace_url" yaml:"workspaceUrl"`
+	// AccessToken is a Databricks AccessToken for a workspace. Information on how
+	// to create tokens can be found here: https://docs.databricks.com/dev-tools/auth.html#personal-access-tokens-for-users
+	AccessToken string `json:"access_token" yaml:"accessToken"`
+	// Databricks needs an Instance Profile with S3 permissions in order to access metadata
+	// storage in S3. Information on how to create this can be found here:
+	// https://docs.databricks.com/aws/iam/instance-profile-tutorial.html
+	S3InstanceProfileARN string `json:"s3_instance_profile_arn" yaml:"s3InstanceProfileArn"`
 }
 
 type EmailConfig struct {
@@ -64,10 +80,6 @@ func (c *EmailConfig) FullHost() string {
 	return fmt.Sprintf("%s:%s", c.Host, c.Port)
 }
 
-type LambdaIntegrationConfig struct {
-	RoleArn string `json:"role_arn" yaml:"roleArn"`
-}
-
 type ConfigBool bool
 
 func (scb *ConfigBool) UnmarshalJSON(data []byte) error {
@@ -86,14 +98,4 @@ func (scb *ConfigBool) UnmarshalJSON(data []byte) error {
 
 	*scb = ConfigBool(b)
 	return nil
-}
-
-type IntegrationConfig map[string]string
-
-func (c *IntegrationConfig) Value() (driver.Value, error) {
-	return utils.ValueJSONB(*c)
-}
-
-func (c *IntegrationConfig) Scan(value interface{}) error {
-	return utils.ScanJSONB(value, c)
 }
